fix(#17): handle case when xdist isn't installed or disabled
Closes: #17
